cleanup legacy v1beta1 policy applier
