+++ pilot/pkg/security/authn/policy_applier_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package v1beta1
+package authn
 
 import (
 	"reflect"
@@ -38,7 +38,6 @@ import (
 	"istio.io/istio/pilot/pkg/features"
 	"istio.io/istio/pilot/pkg/model"
 	"istio.io/istio/pilot/pkg/model/test"
-	"istio.io/istio/pilot/pkg/security/authn"
 	"istio.io/istio/pilot/pkg/util/protoconv"
 	"istio.io/istio/pkg/config"
 	"istio.io/istio/pkg/config/host"
@@ -929,7 +928,7 @@ func TestJwtFilter(t *testing.T) {
 	for _, c := range cases {
 		t.Run(c.name, func(t *testing.T) {
 			istiotest.SetForTest(t, &features.JwksFetchMode, c.jwksFetchMode)
-			if got := NewPolicyApplier("root-namespace", c.in, nil, push).JwtFilter(false, false); !reflect.DeepEqual(c.expected, got) {
+			if got := newPolicyApplier("root-namespace", c.in, nil, push).JwtFilter(false, false); !reflect.DeepEqual(c.expected, got) {
 				t.Errorf("got:\n%s\nwanted:\n%s", spew.Sdump(got), spew.Sdump(c.expected))
 			}
 		})
@@ -1493,7 +1492,7 @@ func TestAuthnFilterConfig(t *testing.T) {
 	}
 	for _, c := range cases {
 		t.Run(c.name, func(t *testing.T) {
-			got := NewPolicyApplier("root-namespace", c.jwtIn, c.peerIn, &model.PushContext{}).AuthNFilter(c.forSidecar)
+			got := newPolicyApplier("root-namespace", c.jwtIn, c.peerIn, &model.PushContext{}).AuthNFilter(c.forSidecar)
 			if !reflect.DeepEqual(c.expected, got) {
 				t.Errorf("got:\n%v\nwanted:\n%v\n", humanReadableAuthnFilterDump(got), humanReadableAuthnFilterDump(c.expected))
 			}
@@ -1573,13 +1572,13 @@ func TestInboundMTLSSettings(t *testing.T) {
 	tlsContextHTTP := proto.Clone(tlsContext).(*tls.DownstreamTlsContext)
 	tlsContextHTTP.CommonTlsContext.AlpnProtocols = []string{"h2", "http/1.1"}
 
-	expectedStrict := authn.MTLSSettings{
+	expectedStrict := MTLSSettings{
 		Port: 8080,
 		Mode: model.MTLSStrict,
 		TCP:  tlsContext,
 		HTTP: tlsContextHTTP,
 	}
-	expectedPermissive := authn.MTLSSettings{
+	expectedPermissive := MTLSSettings{
 		Port: 8080,
 		Mode: model.MTLSPermissive,
 		TCP:  tlsContext,
@@ -1589,7 +1588,7 @@ func TestInboundMTLSSettings(t *testing.T) {
 	cases := []struct {
 		name         string
 		peerPolicies []*config.Config
-		expected     authn.MTLSSettings
+		expected     MTLSSettings
 	}{
 		{
 			name:     "No policy - behave as permissive",
@@ -1606,7 +1605,7 @@ func TestInboundMTLSSettings(t *testing.T) {
 					},
 				},
 			},
-			expected: authn.MTLSSettings{Port: 8080, Mode: model.MTLSDisable},
+			expected: MTLSSettings{Port: 8080, Mode: model.MTLSDisable},
 		},
 		{
 			name: "Single policy - permissive mode",
@@ -1767,11 +1766,11 @@ func TestInboundMTLSSettings(t *testing.T) {
 	}
 	for _, tc := range cases {
 		t.Run(tc.name, func(t *testing.T) {
-			got := NewPolicyApplier("root-namespace", nil, tc.peerPolicies, &model.PushContext{}).InboundMTLSSettings(
+			got := newPolicyApplier("root-namespace", nil, tc.peerPolicies, &model.PushContext{}).InboundMTLSSettings(
 				8080,
 				testNode,
 				[]string{},
-				authn.NoOverride,
+				NoOverride,
 			)
 			if diff := cmp.Diff(tc.expected, got, protocmp.Transform()); diff != "" {
 				t.Errorf("unexpected filter chains: %v", diff)
